# Store timestamp as uint64 in protobuf