--- a/aggregators/internal/timestamppb/timestamppb_test.go
+++ b/aggregators/internal/timestamppb/timestamppb_test.go
@@ -1,0 +1,17 @@
+// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+// or more contributor license agreements. Licensed under the Elastic License 2.0;
+// you may not use this file except in compliance with the Elastic License 2.0.
+
+package timestamppb
+
+import (
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestTimeToPBTimestamp(t *testing.T) {
+	tt := time.Date(2000, 1, 2, 3, 4, 5, 6, time.UTC)
+	assert.Equal(t, tt, PBTimestampToTime(TimeToPBTimestamp(tt)))
+}