Store timestamp as uint64 in protobuf so no pointer is needed for timestamp field. Slight improvement in allocs; sec/op are mostly similar and within error.

benchstat:
```
goos: linux
goarch: amd64
pkg: github.com/elastic/apm-aggregation/aggregators
cpu: 11th Gen Intel(R) Core(TM) i7-11800H @ 2.30GHz
                            │ bench-out/before-proto-timestamp-uint64 │ bench-out/after-proto-timestamp-uint64 │
                            │                 sec/op                  │     sec/op       vs base               │
AggregateCombinedMetrics-16                              4.522µ ± 10%       3.853µ ± 9%  -14.78% (p=0.002 n=6)
AggregateBatchSerial-16                                  20.65µ ± 16%       19.89µ ± 9%        ~ (p=0.240 n=6)
AggregateBatchParallel-16                                22.29µ ±  2%       20.82µ ± 3%   -6.57% (p=0.002 n=6)
CombinedMetricsEncoding-16                               8.585µ ±  3%       8.822µ ± 5%   +2.76% (p=0.041 n=6)
CombinedMetricsDecoding-16                               4.453µ ±  3%       4.505µ ± 3%        ~ (p=0.093 n=6)
CombinedMetricsToBatch-16                                429.1µ ±  2%       432.4µ ± 2%        ~ (p=0.310 n=6)
EventToCombinedMetrics-16                                1.071µ ±  1%       1.107µ ± 3%   +3.36% (p=0.009 n=6)
geomean                                                  12.03µ             11.72µ        -2.62%

                            │ bench-out/before-proto-timestamp-uint64 │ bench-out/after-proto-timestamp-uint64 │
                            │                  B/op                   │       B/op        vs base              │
AggregateCombinedMetrics-16                              3.720Ki ± 1%       3.570Ki ± 4%  -4.04% (p=0.002 n=6)
AggregateBatchSerial-16                                  23.44Ki ± 0%       22.74Ki ± 1%  -2.99% (p=0.002 n=6)
AggregateBatchParallel-16                                23.51Ki ± 0%       22.71Ki ± 0%  -3.41% (p=0.002 n=6)
CombinedMetricsEncoding-16                               2.688Ki ± 0%       2.562Ki ± 0%  -4.65% (p=0.002 n=6)
CombinedMetricsDecoding-16                               10.01Ki ± 0%       10.01Ki ± 0%       ~ (p=1.000 n=6)
geomean                                                  8.878Ki            8.609Ki       -3.03%

                            │ bench-out/before-proto-timestamp-uint64 │ bench-out/after-proto-timestamp-uint64 │
                            │                allocs/op                │   allocs/op     vs base                │
AggregateCombinedMetrics-16                                44.00 ± 2%       41.00 ± 2%  -6.82% (p=0.002 n=6)
AggregateBatchSerial-16                                    195.0 ± 0%       183.0 ± 0%  -6.15% (p=0.002 n=6)
AggregateBatchParallel-16                                  195.0 ± 0%       183.0 ± 0%  -6.15% (p=0.002 n=6)
CombinedMetricsEncoding-16                                 79.00 ± 0%       77.00 ± 0%  -2.53% (p=0.002 n=6)
CombinedMetricsDecoding-16                                 40.00 ± 0%       40.00 ± 0%       ~ (p=1.000 n=6) ¹
geomean                                                    88.03            84.19       -4.37%
¹ all samples are equal
```